Quick bug patch (partially) with some new stuff
The bootloader was unable to load files with a large cluster
index (> 1024) due to the loader only reading the first 8 sectors
of the FAT (1024 clusters; 4 MB). A particular VFAT driver was
placing KERNEL.ELF outside of this range. It's been since updated
to support 2048 cluster index sizes (not ideal but it works for now).

Added partial support for IDT and a simple linked-list memory allocator.